Create Curriculum Responsiveness
Erased Dummy data ng Notifications and History
