+++ src/components/create-curriculum/create-tables.js
@@ -1,126 +1,42 @@
 import {
-  Avatar,
-  Box,
-  Card,
-  Checkbox,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableRow,
-  Typography,
-  Button,
-  CardContent,
-  
-} from '@mui/material';
-import PerfectScrollbar from 'react-perfect-scrollbar';
-  import SaveIcon from '@mui/icons-material/Save';
-  import ArrowBackIcon from '@mui/icons-material/ArrowBack';
-  import List from '@mui/material/List';
-  import ListItem from '@mui/material/ListItem';
-  import ListItemText from '@mui/material/ListItemText';
-  import Divider from '@mui/material/Divider';
-  import TextField from '@mui/material/TextField';
-import { TocTwoTone } from '@mui/icons-material';
-import InputLabel from '@mui/material/InputLabel';
-import FormControl from '@mui/material/FormControl';
-import NativeSelect from '@mui/material/NativeSelect';
-import React, { Component } from 'react';
-
-  const style = {
-    width: '100%',
-    maxWidth: 360,
-    bgcolor: 'background.paper',
-  };
-  const maintext = {
-    color: 'black',
-    fontSize: 20,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  }
-  const summary = {
-    color: 'black',
-    fontSize: 15,
-    width: '100',
-
-  }
-  const when = {
-    color: 'black',
-    fontSize: 13,
-    width: '100',
-  }
-
-  
-
-  function HistoryLog() {
-    return (
-      <List sx={style} 
-      component="nav"
-       aria-label="mailbox folders">
-        <ListItem button>
-          <ListItemText 
-           disableTypography
-           style={maintext}
-           primary="History"/>
-        </ListItem>
-        <Divider />
-        <ListItem button 
-        divider>
-          <ListItemText 
-          disableTypography
-          style={summary}
-          primary="I added a comment"
-          secondary={<div style={when}>User - added 4 days ago</div>}
-          />
-        </ListItem>
-        <ListItem button>
-          <ListItemText 
-          disableTypography
-          style={summary}
-          primary="I added a comment"
-          secondary={<div style={when}>User - added 4 days ago</div>} />
-        </ListItem>
-        <Divider light />
-        <ListItem button>
-          <ListItemText 
-          disableTypography
-          style={summary}
-          primary="I added a comment"
-          secondary={<div style={when}>User - added 4 days ago</div>} />
-        </ListItem>
-        <ListItem button>
-          <ListItemText 
-          disableTypography
-          style={summary}
-          primary="I added a comment"
-          secondary={<div style={when}>User - added 4 days ago</div>} />
-        </ListItem>
-        <ListItem button>
-          <ListItemText 
-          disableTypography
-          style={summary}
-          primary="I added a comment"
-          secondary={<div style={when}>User - added 4 days ago</div>} />
-        </ListItem>
-      
-      </List>
-    );
-  }
-
-  function CreateTables(){
+    Avatar,
+    Box,
+    Card,
+    Checkbox,
+    Table,
+    TableBody,
+    TableCell,
+    TableHead,
+    TableRow,
+    Typography,
+    Button,
+    CardContent,
     
-    return (
-      <Card>
-        <PerfectScrollbar>
-          <Box sx={{ minWidth: 720}}>
-              <TableHead>
-                <TableCell>
+  } from '@mui/material';
+  import PerfectScrollbar from 'react-perfect-scrollbar';
+    import SaveIcon from '@mui/icons-material/Save';
+    import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+    import List from '@mui/material/List';
+    import ListItem from '@mui/material/ListItem';
+    import ListItemText from '@mui/material/ListItemText';
+    import Divider from '@mui/material/Divider';
+    import TextField from '@mui/material/TextField';
+  import { TocTwoTone } from '@mui/icons-material';
+  import InputLabel from '@mui/material/InputLabel';
+  import FormControl from '@mui/material/FormControl';
+  import NativeSelect from '@mui/material/NativeSelect';
+  import React, { Component } from 'react';
+  
+import ArrowRightIcon from '@mui/icons-material/ArrowRight';
+
+
+export const CreateTables = (props) => (
+  <Card {...props}>
+        <Box>
+        <TableHead>
+                <TableCell  sx={{width:'40%' }}>
                 </TableCell>
-                <TableCell>
-               </TableCell>
-                <TableCell >
-                </TableCell>
-                <TableCell>
+                <TableCell sx={{width:'40%'}}>
                     <Box>
                     <FormControl fullWidth>
                     <InputLabel variant="standard" 
@@ -129,7 +45,7 @@ import React, { Component } from 'react';
                       Year Level
                     </InputLabel>
                     <NativeSelect
-                    sx={{ width: 135}}
+                    
                       defaultValue={30}
                       inputProps={{
                         name: 'year',
@@ -144,21 +60,19 @@ import React, { Component } from 'react';
                   </FormControl>
                   </Box>
                 </TableCell>     
-                <TableCell>
+                <TableCell sx={{width:'20%' }}>
                 </TableCell>
-                <TableCell>
-                </TableCell>
-                <TableCell>
-                </TableCell>
-                <TableCell>
+                <TableCell sx={{width:'25%' }}>
                 </TableCell>
               </TableHead>
               <TableCell>
               <p><b>First Semester</b></p>
               </TableCell>
-
- {/*First Semester Header*/}
-              <TableBody>
+  
+    <Divider />
+    
+    {/*First Semester Header*/}
+    <TableBody>
               <TableCell sx={{fontWeight: 'bold' }}>
                     COURSE CODE
                   </TableCell>
@@ -187,10 +101,7 @@ import React, { Component } from 'react';
                   </TableBody>
 
   {/*First Semester TextFields*/}  
-
-                  
-                   
-                   <Table label =" Line1">
+     <Table label =" Line1">
             <TableCell>
                    <TextField sx={{ minWidth: 95 }}
                     id="standard-multiline-static"
@@ -1756,6 +1667,7 @@ import React, { Component } from 'react';
               >
               <b>0</b>
               </TableCell>
+           
               <TableCell>
   
               </TableCell>
@@ -1765,98 +1677,8 @@ import React, { Component } from 'react';
               <TableCell>
               </TableCell>
             </Table>
+       
 
-                    
-
-                  
-            
-          </Box>
-        </PerfectScrollbar>
-      </Card>
-      
-    );
-  }
-  export const CreateCurriculumListToolbar = (props) => (
-    <Box {...props}>
-      <Box
-        sx={{
-          alignItems: 'center',
-          display: 'flex',
-          justifyContent: 'space-between',
-          flexWrap: 'wrap',
-          m: -1,
-  
-        }}
-      >
-        <Typography
-          sx={{ m: 1 }}
-          variant="h4"
-        >
-         Create Curriculum
-        </Typography>
-        <Box sx={{ m: 1 }}>
-          <Button
-            startIcon={(<ArrowBackIcon fontSize="small" />)}
-            sx={{ mr: 1 }}
-          >
-            Cancel
-          </Button>
-          <Button
-            color="primary"
-            variant="contained"
-            startIcon={(<SaveIcon fontSize="small" />)}
-          >
-            Save
-          </Button>
-        </Box>
-      </Box>
-      <Box sx={{
-         mt: 3 
-         }}>
-           
-        <Card sx={{
-          flexDirection: 'row',
-          display: 'flex',
-          p:1,
-          m:1,
-
-        }}>
-  
-          <CardContent
-          sx={{
-            width:'100%',
-          }}
-          >
-            <Box sx={{ 
-              width: '80%', 
-              backgroundColor: 'light gray',
-              border: 'black',
-              height: '100%',
-              borderRadius: 1,
-              border: '1px solid #D3D3D3',
-              
-              }}>
-                <HistoryLog/>
-            </Box>
-          </CardContent>
-          <CardContent>
-          <Box sx={{ 
-              width: '100%', 
-              backgroundColor: 'light gray',
-              border: 'black',
-              height: '100%',
-              borderRadius: 1,
-              border: '1px solid #D3D3D3'
-             
-              }}>
-            <CreateTables/> 
-            </Box>
-      
-          </CardContent>
-          
-        
-        </Card>
-      </Box>
     </Box>
-  );
-  
+  </Card>
+);
