+++ src/pages/NotificationList.js
@@ -47,36 +47,11 @@ function NotificationDiv(props){
   }}
   >
     <ListItem>
-      <ListItemAvatar>
-        <Avatar>
-          <ImageIcon />
-        </Avatar>
-      </ListItemAvatar>
-      <ListItemText primary="Photos" 
-      secondary="Jan 9, 2014" />
+      <ListItemText primary=" No Notifications Yet" 
+      secondary=" " />
     </ListItem>
     <Divider variant="inset" 
     component="li" />
-    <ListItem>
-      <ListItemAvatar>
-        <Avatar>
-          <WorkIcon />
-        </Avatar>
-      </ListItemAvatar>
-      <ListItemText primary="Work" 
-      secondary="Jan 7, 2014" />
-    </ListItem>
-    <Divider variant="inset"
-     component="li" />
-    <ListItem>
-      <ListItemAvatar>
-        <Avatar>
-          <BeachAccessIcon />
-        </Avatar>
-      </ListItemAvatar>
-      <ListItemText primary="Vacation" 
-      secondary="July 20, 2014" />
-    </ListItem>
     <ListItem>
     <NextLink
       href="/notifications"
